+++ node/resolvers/Mutations/Users.ts
@@ -379,7 +379,7 @@ const Users = {
     ctx: Context
   ) => {
     const {
-      clients: { storefrontPermissions: storefrontPermissionsClient },
+      clients: { events, storefrontPermissions: storefrontPermissionsClient },
       vtex: { adminUserAuthToken, logger, sessionData, storefrontPermissions },
     } = ctx as Context | any
 
@@ -413,6 +413,11 @@ const Users = {
         userId,
       })
       .then((result: any) => {
+        events.sendEvent('', 'b2b-organizations-graphql.removeUser', {
+          id,
+          email,
+        })
+
         return result.data.deleteUser
       })
       .catch((error: any) => {