+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.39.0] - 2023-09-25
+
 ### Added
 
 - Add metrics to maintain buyer organization users